requirements_gpu.txt
blinker==1.8.1
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
fsspec==2024.2.0
h5py==3.11.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sympy==1.12
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
typing_extensions==4.9.0
Werkzeug==3.0.2
